Use normal for loop instead of for/in
For/in, at least in our environment, winds up iterating over the
object's prototype methods such as `_super` and `nextObject`. These
result in a _bunch_ of `class="emoji-undefined"` elements being added to
the UI which makes the sections scroll way farther than needed.

Switching for/in to a normal for loop because these are just arrays so
why bother with for/in and hasOwnProperty?